services: Register lmo face unlock as weak biometric
Our face unlock doesn't meet the requirements for
Class 3 (Strong) biometric so whenever third party apps
tries to access the keystore with face auth, it crashes.

Hence, mark it as weak authenticator and so it can't be
used in third party apps that requires strong authentication.

Also we don't recommend to use face auth for those usecases.

Change-Id: I66afd0126582f008c18c3f504094837f062b9b44